Move 3D Objects into own package
